Update spinner component name to TouchSpin, add new component spinner
