+++ src/main/java/net/bootsfaces/component/ComponentsEnum.java
@@ -60,13 +60,14 @@ public enum ComponentsEnum {
 	scrollSpy("<b:scrollSpy", "scrollSpy"),
 	selectBooleanCheckbox("<b:selectBooleanCheckbox", "selectBooleanCheckbox"),
 	shake("<b:shake", "shake"),
+	spinner("<b:spinner", "spinner"),
 	switchComponent("<b:switch", "switch"),
 	selectMultiMenu("<b:selectMultiMenu", "selectMultiMenu"),
 	selectOneMenu("<b:selectOneMenu", "selectOneMenu"),
 	slider("<b:slider", "slider"),
 	slider2("<b:slider2", "slider2"),
 	socialShare("<b:socialShare", "socialShare"),
-	spinner("<b:spinner", "spinner"),
+	touchSpin("<b:touchSpin", "touchSpin"),
 	tab("<b:tab", "tab"),
 	tabLinks("<b:tabLinks", "tabLinks"),
 	tabView("<b:tabView", "tabView"),
