+++ src/main/java/net/bootsfaces/component/kebab/Kebab.java
@@ -19,10 +19,9 @@
 package net.bootsfaces.component.kebab;
 
 import javax.el.ValueExpression;
+import javax.faces.component.FacesComponent;
+
 import net.bootsfaces.listeners.AddResourcesListener;
-import javax.faces.application.ResourceDependencies;
-import javax.faces.application.ResourceDependency;
-import javax.faces.component.*;
 import net.bootsfaces.utils.BsfUtils;
 
 /** This class holds the attributes of &lt;b:kebab /&gt;. */
