+++ src/main/java/net/bootsfaces/component/kebab/KebabRenderer.java
@@ -18,16 +18,14 @@
 
 package net.bootsfaces.component.kebab;
 
-import javax.faces.component.*;
 import java.io.IOException;
-import java.util.Map;
 
+import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import javax.faces.context.ResponseWriter;
 import javax.faces.render.FacesRenderer;
 
 import net.bootsfaces.render.CoreRenderer;
-import net.bootsfaces.render.Tooltip;
 
 /** This class generates the HTML code of &lt;b:kebab /&gt;. */
 @FacesRenderer(componentFamily = "net.bootsfaces.component", rendererType = "net.bootsfaces.component.kebab.Kebab")
